Fix checks for contexts being inside Providers
The checks for TableDataProvider and VariablesProvider did not work,
since both have default variables. This expands them to have an
'initialized' variable that defaults to false, and then is set to true
when creating the context.

Also had to fix some tests, since they now correctly were failing. And
also some had to fix some unhandled errors in fetchTableData.