Fix missing type-module in package.json files
This adds the line '"type": "module"' to the package.json files that are
missing it as of now. It cleans up the messages we get when we build the
projects. There were several places where it complained that we
built it using the old "require()" way. This made reading the terminal
output more difficult than it needs to be.

It also adds some formatting to one vite.config.ts file, but nothing
functional in it has changed.