Remove the last types from NX in the projecct
This fixes the errors we have in the tsconfig files from the old types
from NX that were left. After this there should not be many more times
where we have the red 'error' indicator in our IDEs from config files.
One exception is the one file with a test part, that causes an error.