Add mocking in test for useTranslate (#493)
* Add mocking in test for useTranslate

This fixes the spam we have had in the terminal related to react-i18next
and the useTranslate function. It has been quite spammy, making the
testing results hard to read.

Took some troubleshooting to get it to work, but made a mocking function in
the ui library, and the export it to the web app to use it there too.
The lines for importing and using them are at the top of the files,
because they need to be ran before importing the other components that
use the function. This seems clearer and easier to deal with than trying
to figure out which import line needs to be moved below the others.

Also fixed some small issues with missing variables in the tests where I
saw them.

* Add sonarcloud config file (trying it)

Since we have ran into the issue that sonarcloud several times complains
about duplicate code, in files we think it makes sense in, (test files,
two similar components), we want to try and see if we can make a config
to add exlusions to its analysis in the pipeline.

* Refactor mock for less nesting of functions