Fix PxWeb2 font-family and font-weights
This fixes how we create our own font-family in the project. The
font-family was initially created so we had 3, one for each font-weight.
This leads to issues with browser user-agents, that automatically try to
add boldness etc. to h1 tags and similar. For us, all the titles were
too bold. Now the font-families are merged into one, with the
font-weights: 400, 500 and 700. The site looks the same in chrome and
safari.