Add keyboard nav and focus to the Select in VarBox
This adds the keyboard navigation and focus marking to the Select and
modal in the variableBox. It reforcuses the Select when the modal
closes.

What is left is to look into a solution for the Enter key when closing
the modal, since that does not refocus correctly. Maybe there is
something special about the Enter key and how browsers handle it.