Refactor and add handling for Escape key presses
Since the component should be usable in other projects, it needs to
handle escape key presses also. It can not rely on a provider in the
frontend application for this.

Since we now add an event listener for keydown events, overriding the
Enter keys default behaviour on buttons also done in there, instead of
on each button. Should work fine, for now atleast.