Add keyboard nav and focus to the Select in VarBox (#535)
This adds the keyboard navigation and focus marking to the Select and Modal in the variableBox. It refocuses the Select when the modal closes. It handles keypresses according to the design in Figma.

It also expands the test suite for the Modal, Radio and Select components. We should now test for
more of the core functionality of them.