Fix/px-web2-568: search focusmarking and keyboard nav (#553)
Adds the correct keyboard navigation and focusmarking for the search component. It also greatly expands on the unit tests and the Stories in Storybook, which now more accurately reflects what the component can do.

It fixes the focus management when canceling a search. And handles both escape when input is focused, and enter, spacebar and escape when cancel button is focused/pressed. It also makes sure the search does not completely overwrite the keyboard handling/navigation. For example, it allows the normal Escape functionality to work in the PxWeb2 app, when there is no active search value. Allowing the Escape key to close the NavigationBar/Rail.

It also fixes a visual bug with the search in the Variablebox, in regards to the border-radius/corners.