Add new global and semantic colors
This adds the new global and semantic colors according to PxWeb2-460. It
refactors the highlight component into it's own files in the checkbox
component, for ease of use and readability. Also makes it easy to move
it later if wanted.

This does not fix the issues discovered with the existing colors yet.