Update Storybook: 8.5.8 -> 8.6.4
This update the storybook packages to the latest versions. Ran the
Storybook upgrade command, so we would get any migrations. Storybook
recommends using 'react-docgen' since it is faster, but it is not
accurate enough for our usecase. Union types were shown as [union],
instead of the different values. Changed it back to
'react-docgen-typescript'.