Feature/pxweb2-460: Add global color values and fix semantic colors in the default theme (#567)
This adds the new global and semantic colors according to PxWeb2-460. It refactors the highlight component into it's own files in the checkbox component, for ease of use and readability. And makes it easy to move it later if wanted.

The PR also fixes the missing global colors, and makes all semantic colors reference the correct global color codes.

Since we want to easily be able to read and use the theme file, this adds sections like "background" instead of each semantic color being on its own line with a "background-" prefix. It now also matches how they are in Figma from the designer.

Also removed some unused semantic colors.

* Fix key in Highlight Component