Update NavigationBar/Rail for keyboard navigation
This updates the NavigationBar/Rail to handle keyboard navigation
without needing any tabIndex changes. The components and app.tsx were
refactored so the HTML barely changes when switching between the
horizontal and vertical versions. Also added a check for if the
SkipToMain is focused, so it can be used in calculating heights for
elements.

Also, we might consider refactoring the NavigationRail/Bar into one
component, since the functionality is almost 100% the same. But have
not done that as of now, since it might be a bigger jobb in fixing the
CSS for that.