Fix/install specific versions of react typescript (#591)
This installs specific versions of React and it's types, since we want
to be able to control when we update them.

It also install a specific Typescript version in the root package.json.
Since we want to make sure we control which version of Typescript we are
using. And it removes the tech-debt task about this from the Maintainers.md

Lastly, it fixes a type error in VariableBoxContent after installing the
latest version of TypeScript.