Add maxDataCells to web app config
Since we need to have a number that indicates the maximum number of
cells we support in the TableView component, and the api-calls it makes.
We need to add a field to the web app config that can be configured and
used for this purpose.

Later we might add functionality so that this setting can be directly
gotten from the api also/instead.