fix: Variable's values' txts when changing codelists (#615)
* Fix the variable's values' txts changing codelists

Since we want the ability to change codelists, we need to update the
function that parses the new codelist metadata for a variable to
properly handle the ValueDisplayType and value label texts. This adds
similar parsing as the Json2Mapper.

Small refactor to extract shared functionality. If we decide to
redo/change how we get the metadata when changing codelists, some of
this might have to be redone/undone again.

* Refactor getLabelText to remove 'else'