Update Chevron in Select to point sideways
Since the downwards chevron can be a bit misleading, we are changing it
to be a sideways chevron in the 'inVariableBox' version. To better
indicate that something will open in the middle of the screen.

This contains some refactoring, since some functionality is now shared.
Added a function to show 2 different icons based on the language
direction, but not sure if this is what we want. The original code for
this check is one line, so maybe we just want to use that instead.
Would like feedback on this!

Also added a 'RTL' part to the Select in Storybook, so we can see how
it looks when the direction is reversed!