Add language to url for language persistance
Since we need to support several languages with the project, and we want
people to have the ability to share links, we need to add language
persistance. What complicates this is that a table can have only one
language, even though several can be supported for the entire site. So
we need it to work when people share links with other people.

This adds an optional language part to the url, and it uses the fallback
language if that part is missing. This means that urls like
'/sv/table/tableid' and '/table/tableid' both will work. Since it seems
unlikely a site will change its default/fallback language.

I installed the i18next package for detecting browser languages, but set
it up to only check for path, and to not cache the values. Since we only
want the language of the URL to matter.

This has not added any tests yet, but I had to update and change some
tests for the ErrorPage and the ErrorBoundry, since those broke with
how the LanguageSwitcher works, or were spamming the output of the test
runs.