Add config for showing the default lang in URL
Since different places that will be using the PxWeb2 project might have
different requirements for the language in the URL, we need to support
both implementations. This adds the config and routing to support both
showing or hiding the default language in the URL:
    1. Always showing the language in the URL: '/en/table'
    2. Only showing the non-default languages in the URL: '/table'

This also adds a catch-all route in the routing, for when a given URL
matches no routes. Currently it only displays an empty div, but we will
need a NotFound/404 component here.

There are 2 redirects left in the routing, but these should be removed
later, when they are no longer needed for development.