Feature/px web2 432 add support for always default lang in url (#688)
Since we need to support several languages with the project, and we want
people to have the ability to share links, we need to add language
persistance. What complicates this is that a table can have only one
language, even though several can be supported for the entire site. So
we need it to work when people share links with other people.

This adds an optional language part to the url, and it uses the fallback
language if that part is missing. This means that urls like
'/sv/table/tableid' and '/table/tableid' both will work. Since it seems
unlikely a site will change its default/fallback language.

I installed the i18next package for detecting browser languages, but set
it up to only check for path, and to not cache the values. Since we only
want the language of the URL to matter.

* Fix type error in Selection

* Refactor path logic to utility function

* Add and refactor tests

This adds and refactors tests for ErrorPage, LanguageSwitcher and the
utility function getLanguagePath.

Used Claude 3.7 to generate tests and worked from there.

* Add config for showing the default lang in URL

Since different places that will be using the PxWeb2 project might have
different requirements for the language in the URL, we need to support
both implementations. This adds the config and routing to support both
showing or hiding the default language in the URL:
    1. Always showing the language in the URL: '/en/table'
    2. Only showing the non-default languages in the URL: '/table'

This also adds a catch-all route in the routing, for when a given URL
matches no routes. Currently it only displays an empty div, but we will
need a NotFound/404 component here.

* Fix routing for both URL types

This fixes the routing for URLs with and without the default language.
The only thing missing now is the redirect for URLs without a language,
when the default language should be in the URL.

* Add NotFound component and move router logic to it's own file

* Add basic tests to NotFound component

* Re-add redirect to root path when language should be in the URL

* Add working tests for the routing in the web app

This adds the tests for the new routing logic.

* Make NotFound props readonly

* Change to page instead of table not found

* Fix bug in LanguageSwitcher sometimes showing wrong language

* Update swedish languagefile

* Fix url in startpage to go to tableview