Add proper title handling to NavigationDrawer (#712)
* Add proper title handling to NavigationDrawer

Since we are starting to work on the different pages/views of the
sidebar/NavigationDrawer component, we need to fix how it currently
always displays the title for the filter view. Even on other views.

This adds the functionality for handling that. It needed some
renaming for ease of use in the future, and to bring the NavDrawer
closer to the translation keys.

* Remove unnecessary type assertion (according to sonarqube)