Refactor to memoize more when drawing the Table
Refactor to minimize the rerendering of the Table component, either when
loading data or pivoting it. When testing the pivot on a large Table
(close to 150000 cells), it could take 4+ seconds. This should
significantly improve that by memoizing more, and also by moving the
Navigation Drawer's views into seperate components. Did not move the
VariableList yet.

Co-authored-by: Mikael Nordberg <54406826+MikaelNordberg@users.noreply.github.com>