Add caching to cell strings formatting
Since we want to minimize the time it takes to render tables, we should
only do the formatting with react-i18next when needed. And we should
cache the translation/formatting.

This expands on the formatting, so we now can cache after the first
time. During some testing the time to format the values in the table
could take 300+ ms, and this caching removes that step in the table
rendering after the first time the value is shown.

Also added tests.