Add pivot clockwise to navigation drawer edit pane (#714)
* Add button under Edit view to call TableDataProvider's pivotCW function

* Add dynamic title to ContentTop for selection

Adds a dynamic title based on the current selection, to the ContentTop

* Remove dev console.logs and fix test

* Fix formatting in resources.d.ts

* Small refactor of variables and imports

* Refactor to memoize more when drawing the Table

Refactor to minimize the rerendering of the Table component, either when
loading data or pivoting it. When testing the pivot on a large Table
(close to 150000 cells), it could take 4+ seconds. This should
significantly improve that by memoizing more, and also by moving the
Navigation Drawer's views into seperate components. Did not move the
VariableList yet.

Co-authored-by: Mikael Nordberg <54406826+MikaelNordberg@users.noreply.github.com>

* Consolidate imports

* Add caching to cell strings formatting

Since we want to minimize the time it takes to render tables, we should
only do the formatting with react-i18next when needed. And we should
cache the translation/formatting.

This expands on the formatting, so we now can cache after the first
time. During some testing the time to format the values in the table
could take 300+ ms, and this caching removes that step in the table
rendering after the first time the value is shown.

Also added tests.

* Update Table fakeData function for storybook

* Move formatting to utils file and update tests

* Update DrawerSave test

* Move formatting to fetchFromApi

* Add support for 6 decimals when formatting

---------

Co-authored-by: Mikael Nordberg <54406826+MikaelNordberg@users.noreply.github.com>