Change when NavDrawer is open by default (#767)
This changes when the NavDrawer is opened by default on initial table
page load. Since we found issues with focus being hidden behind it, it
was decided that a better solution for keyboard users is to not start in
a state where focus can be hidden. Now only shows it by default on
larger than the large breakpoint.

Considered adding a simpler check in TableViewer, but decided to expand
the solution for checking window width in the AppProvider. To be
consistent with previous solutions.