Make PivotButton read texts on focus and click
This expands the PivotButton component in DrawerEdit, so it gives two
different texts to screen readers to read. The first one if the text of
the button, when it is focused. The second texts is the new, order of
the variables, like they currently are shown in the title of the table.