Fix DrawerSave missing labels and expand tests
Fix missing labels in the DrawerSave component, due to the ActionItem
implementation having changed. Also expanded tests for the save to file
functionality of the DrawerSave component, since there were no tests
for that, only other functionality inside the DrawerSave component.

Used GPT-5 (preview) in CoPilot to help and generate much of this.
However, decided to not implement some tests for more flakey
implementations, like the "more than 5 seconds" announcement.