Change the PivotButton to have proper design (#802)
* Change the PivotButton to have proper design

This changes the PivotButton to have the correct design and
aria functionality. Refactored the ActionItem slightly, so it can have
different label and aria-label.

* Make PivotButton read texts on focus and click

This expands the PivotButton component in DrawerEdit, so it gives two
different texts to screen readers to read. The first one if the text of
the button, when it is focused. The second texts is the new, order of
the variables, like they currently are shown in the title of the table.

* Fix test output spam due to state outside act

* Add both aria-label and screen reader texts Pivot

Add so the PivotButton has 3 texts from translations.

1. The button text that is shown
2. The aria-label text that screen readers read on focusing the button
3. The screen reader announcement that is read after pivoting

* Change to using output instead of span for screenreader content

* Fix DrawerSave missing labels and expand tests

Fix missing labels in the DrawerSave component, due to the ActionItem
implementation having changed. Also expanded tests for the save to file
functionality of the DrawerSave component, since there were no tests
for that, only other functionality inside the DrawerSave component.

Used GPT-5 (preview) in CoPilot to help and generate much of this.
However, decided to not implement some tests for more flakey
implementations, like the "more than 5 seconds" announcement.