I18next should use default lang
This fixes the logic for the language system, so it uses the default
language when no language match is found in the URL. It is needed when
the default language should not be in the URL.

It does this by adding a new, custom detector that always returns the
default language, from the config file. This is put behind the URL
matching, so it will only be used when no language matches.