Refactor aria hidden and label to be exclusive (#973)
* Refactor aria hidden and label to be exclusive

This changes the logic for the different icons to handle the aria-label
and aria-hidden exclusivly; they can no longer have both at the same
time. The BreadcrumbIcon is always set to aria-hidden true, since they are
always decorative.

Also expand on the tests for the Icon components.

* Remove aria-label from search icon in variablebox

The search in the variablebox had an aria-label on the search icon, but
this icon is decorative. This removes that label and updates the test
and llanguage files.