Update node moduleResolution version (#991)
* Update node moduleResolution version

We are using an old version of node moduleResolution, that is
deprecated. "node" should be "node10" according to the docs. However,
instead this updates the node version here to the newest, nodenext.

* Change to moduleResolution bundler instead

The bundler option is the more correct option for us, since we use Vite.

Also had to change the type target for the api client. Not sure why it
was set to CommonJS, but don't think it uses any CommonJS specific
patterns.

* Remove 'module' option from tsconfig.json

Removed 'module' option from compiler options in the api-client.